# artifacts/pos-system/src/backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant ERP + POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productNameAr:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, preparing, ready, served]
      required:
        - id
        - productId
        - productName
        - productNameAr
        - quantity
        - unitPrice
        - totalPrice
        - status
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        type:
          type: string
          enum: [dine_in, takeaway, delivery]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, served, paid, cancelled]
        tableId:
          type: ["integer", "null"]
        tableName:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        paymentMethod:
          type: ["string", "null"]
          enum: [cash, card, transfer, wallet, null]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - type
        - status
        - items
        - subtotal
        - discount
        - tax
        - total
        - createdAt
        - updatedAt
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        notes:
          type: ["string", "null"]
      required:
        - productId
        - quantity
    OrderInput:
      type: object
      properties:
        type:
          type: string
          enum: [dine_in, takeaway, delivery]
        tableId:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        discount:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - type
        - items
    OrderUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [dine_in, takeaway, delivery]
        tableId:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        discount:
          type: number
        notes:
          type: ["string", "null"]
...